agent/requirements.txt
requests
psutil
polars
torch
numpy
wandb
python-dotenv
